mysocket/requirements.txt
Click==7.0
h11==0.8.1
httptools==0.0.13
python-engineio==3.10.0
python-socketio==4.4.0
six==1.11.0
uvicorn==0.7.0
uvloop==0.12.2
websockets==7.0
